test bug
git-svn-id: https://timeseriesdb.googlecode.com/svn/trunk@22 537f4964-0365-11de-bcc6-d9108be23e13

+++ FastBinTimeseries.Test/BinIndexedFileCoreTests.cs
@@ -165,9 +165,11 @@ namespace NYurik.FastBinTimeseries.Test
                 {
                 }
 
-                File.Delete(fileName);
-
-                f.InitializeNewFile();
+                if (RunMode == Mode.OneTime)
+                {
+                    File.Delete(fileName);
+                    f.InitializeNewFile();
+                }
             }
         }
     }
